# excel/backend/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "This is for SMARTCARE PLUS  - A REST API for SMARTCARE PLUS ",
    "version": "1.0.6",
    "title": "SMARTCARE PLUS API"
  },
  "host": "localhost:3050",
  "basePath": "/api/v1",
...
  "schemes": ["http", "https"],
  "paths": {
...
    "/users/update-me": {
      "patch": {
        "tags": ["Authentication"],
        "summary": "update login user",
        "description": "Anyone can access this route to signup",
        "operationId": "updateMe",
        "consumes": ["multipart/form-data"],
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "firstName",
            "in": "formData",
            "required": false,
            "type": "string"
          },
          {
            "name": "lastName",
            "in": "formData",
            "required": false,
            "type": "string"
          },
          {
            "name": "email",
            "in": "formData",
            "required": false,
            "type": "string"
          },
          {
            "name": "contactNumber",
            "in": "formData",
            "required": false,
            "type": "string"
          },
          {
            "name": "gender",
            "in": "formData",
            "required": false,
            "type": "string",
            "enum": ["male", "female", "other"]
          },
          {
            "name": "photo",
            "in": "formData",
            "required": false,
            "type": "file"
          }
        ],
        "responses": {
          "default": {
            "description": "successful operation"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
...
    "/users/delete-me": {
      "delete": {
        "tags": ["Authentication"],
        "summary": "delete login user",
        "description": "Anyone can access this route to signup",
        "operationId": "deleteMe",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [],
        "responses": {
          "default": {
            "description": "successful operation"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/users/approve/{id}": {
      "get": {
        "tags": ["Users"],
        "summary": "approve user",
        "description": "Anyone can access this route to approve user",
        "operationId": "approveUser",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "User id to approve",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "default": {
            "description": "successful operation"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    }
  },
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  },
...
}